Update prompt, move prompt generator to separate file
